indexer/requirements.txt
minio==7.2.7
psycopg2-binary==2.9.9
duckdb==0.10.3
pydantic==2.7.1
pydantic-settings==2.3.0
